csp-server/requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
DateTime==6.0
Flask==3.1.2
Flask-Scss==0.5
idna==3.11
influxdb-client==1.49.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.2
paho-mqtt==2.1.0
pyScss==1.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
reactivex==4.1.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
zope.interface==8.1.1
